pyproject.toml
[tool.poetry]
name = "awsomeHypixelBot"
version = "0.1.0"
description = "awsomeHypixelBot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"
hypixel = "^0.8.0"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = []
plugin_dirs = ["awsomehypixelbot/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
